Added and implemented env var EMBEDDING_TIMEOUT
